% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_wgs84.R
\name{acc_wgs84}
\alias{acc_wgs84}
\title{Calculate total acceleration}
\usage{
acc_wgs84(latitude)
}
\arguments{
\item{latitude}{The latitude in degrees.}
}
\value{
g given in units of $m/s^2$
}
\description{
This function calculates the total acceleration due to gravitation and centripetal force at the earth's surface according to the WGS84 international gravity formula.
}
\note{
Source: http://solid_earth.ou.edu/notes/potential/igf.htm
}
\examples{
acc_wgs84(50)
# Returns: 9.8107 $m/s^2$
}
