% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beaked_whale}
\alias{beaked_whale}
\title{Set of sensor lists for a beaked_whale}
\format{A set of sensor lists:
\describe{
  \item{A}{sensor list contining a triaxial acceleration matrix sampled at 25 Hz}
  \item{M}{sensor list containing a triaxial magnetometer matrix sampled at 25 Hz}
  \item{P}{sensor list containing a pressure (depth) vector sampled at 25 Hz}
}}
\usage{
beaked_whale
}
\description{
Data is from a _Mesoplodon densirostris_ with tag ID md13_134a. The device used was a DTAG3 and it was deployed at 2013-05-14 12:42:00 in El Hierro, Canary Islands, Spain.
}
\keyword{datasets}
