% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_no_delay.R
\name{buffer_no_delay}
\alias{buffer_no_delay}
\title{Buffers a signal vector into matrix of data frames with no delay.}
\usage{
buffer_no_delay(vec, n, p)
}
\arguments{
\item{vec}{The signal vector to be buffered}

\item{n}{The desired length of data segments (rows).}

\item{p}{The desired amount of overlap between consecutive frames (columns) in the output matrix}
}
\value{
A matrix of the buffered signal vector "vec" with "n" data segments and an overlap between consecutive frames specified by "p".
}
\description{
Buffers a signal vector into matrix of data frames with no delay.
}
