% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_opt.R
\name{buffer_opt}
\alias{buffer_opt}
\title{Buffers a signal}
\usage{
buffer_opt(x, n, p, opt)
}
\arguments{
\item{x}{The signal vector to be buffered}

\item{n}{The desired length of data segments (rows).}

\item{p}{The desired amount of overlap between consecutive frames (columns) in the output matrix}

\item{opt}{The vector of samples specified to precede x[1] in an overlapping buffer}
}
\value{
X A matrix of the buffered signal vector "vec" with "n" data segments and an overlap between consecutive frames specified by "p". The matrix starts with "opt" values.

z The remainder of the vector which was not included in the matrix if the last column did not have a full number of rows.

opt The last values, length of "p", of the matrix "X".
}
\description{
This function is used to buffer a signal vector into matrix of data frames and specifies a vector of samples to precede x[1] in an overlapping buffer.
}
