% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_line3.R
\name{col_line3}
\alias{col_line3}
\title{Plot coloured line(s) in 3 dimensions}
\usage{
col_line3(x, y, z = 0, c)
}
\arguments{
\item{x}{a vector or matrix of points on the horizontal (x) axis.}

\item{y}{A vector or matrix of points on the vertical (y) axis.}

\item{z}{A vector or matrix of points on the third (z) axis.}

\item{c}{A vector or matrix of values representing the colour to draw at each point.}

\item{col_lab}{A string to use as the label for the color legend}

\item{...}{Additional inputs for plot_ly()}
}
\description{
This function is used to plot three dimensional lines with each individual line possessing a different color.
}
\note{
x, y, z and c must all be the same size vectors. The color axis will by default span the range of values in c, i.e., caxis will be c(min(min(c)), max(max(c))).
}
\examples{
col_line3(1:20, 1:20, 1:20, 1:20)
}
\seealso{
\code{\link{col_line}}, \code{\link{cline}}
}
