% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conv_cpp}
\alias{conv_cpp}
\title{Convolution (in C++)}
\usage{
conv_cpp(a, b, shape = "full")
}
\arguments{
\item{a}{The first numeric vector.}

\item{b}{The second numeric vector.}

\item{shape}{(optional) The shape of the output: "full", "same", or "valid". Default is "full," matching \code{\link[signal]{conv}} and the Armadillo default, which means that the output array is \code{length(a)} + \code{length(b)} - 1.}
}
\value{
A numeric vector containing the convolution result.
}
\description{
Performs 1D convolution of two vectors in C++ via FFT.
}
