% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cst2hms.R
\name{cst2hms}
\alias{cst2hms}
\title{Convert a time to year, month, day, hours, minutes, seconds local time.}
\usage{
cst2hms(tag, cst, d3, TAGON, GMT2LOC)
}
\arguments{
\item{tag}{The tag id string (eg "zc11_267a")}

\item{cst}{A scalar or a vector of times to convert (in seconds since start of recording)}

\item{d3}{Give the value 0 if the tag was a dtag2, or 1 if a d3 (if 1, then it will be assumed that the TAGON in the cal file is in UTC and GMT2LOC will be used to covert to local).}

\item{TAGON}{A vector with tagon time as [yyyy mm dd hh mm ss], in case there is no cal file for this tag available.}

\item{GMT2LOC}{(optional) The conversion factor for GMT to local time, if d3=1 and tagon is given then GMT2LOC should be given as well.}
}
\value{
hms The output for a date/time string which is a matrix of strings where row n is a string indicating the local date and time for entry n of csts.
}
\description{
This function is used to convert a time in CST (seconds since start of recording) to year, month, day, hours, minutes, seconds local time.
}
