% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_rate.R
\name{depth_rate}
\alias{depth_rate}
\title{Estimate the vertical velocity by differentiating a depth or altitude time series. A low-pass filter reduces the sensor noise that is amplified by the differentiation.}
\usage{
depth_rate(p, fs, fc)
}
\arguments{
\item{p}{A depth or altitude time series or list of depth or altitude series (e.g., from readtag.R). p can have any units and is in the form of a vector}

\item{fs}{is the sampling rate of p in Hz.}

\item{fc}{(optional) A smoothing filter cut-off frequency in Hz. If fc is not given, a default value is used of 0.2 Hz (5 second time constant).}
}
\value{
v The vertical velocity with the same sampling rate as p. v has the same dimensions as p. The unit of v depends on the unit of p. If p is in meters, v is in meters/second
}
\description{
Estimate the vertical velocity by differentiating a depth or altitude time series. A low-pass filter reduces the sensor noise that is amplified by the differentiation.
}
\note{
The low-pass filter is a symmetric FIR with length 4fs/fc. The group delay of the filters is removed.
}
