% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dives.R
\name{find_dives}
\alias{find_dives}
\title{Find time cues for the start and end of either dives in a depth record or flights in an altitude record.}
\usage{
find_dives(p, fs, mindepth, surface = NULL, findall = NULL)
}
\arguments{
\item{p}{A depth or altitude time series (vector) in meters.}

\item{fs}{The sampling rate of the sensor data in Hz (samples per second).}

\item{mindepth}{The threshold in meters at which to recognize a dive or flight. Dives shallow or flights lower than mindepth will be ignored.}

\item{surface}{(optional) The threshold in meters at which the animal is presumed to have reached the surface. Default value is 1. A smaller value can be used if the dive/altitude data are very accurate and you need to detect shallow dives/flights.}

\item{findall}{(optional) When 1 forces the algorithm to include incomplete dives at the start and end of the record. Default is 0 which only recognizes complete dives.}
}
\value{
T is a structure array with size equal to the number of dives/flights found. The fields of T are: start (time in seconds of the start of each dive/flight), end (time in seconds of the start of each dive/flight), max (maximum depth/altitude reached in each dive/flight), tmax	(time in seconds at which the animal reaches the max depth/altitude). If there are n dives/flights beyond mindepth in p, then T will be a structure containing n-element vectors.
}
\description{
Find time cues for the start and end of either dives in a depth record or flights in an altitude record.
}
