% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_researcher.R
\name{get_researcher}
\alias{get_researcher}
\title{Find matching researcher in a list of known tag researchers}
\usage{
get_researcher(initial)
}
\arguments{
\item{initial}{a two-letter code for the species of interest (first letter of Genus and first letter of species)}
}
\description{
Find matching researcher in a list of known tag researchers
}
