% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{Find matching species in a list of marine mammals}
\usage{
get_species(initial)
}
\arguments{
\item{initial}{a two-letter code for the species of interest (first letter of Genus and first letter of species)}
}
\description{
Find matching species in a list of marine mammals
}
