% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harbor_seal}
\alias{harbor_seal}
\title{Set of sensor lists for a harbor seal}
\format{
A set of sensor lists:
\describe{
  \item{A}{sensor list contining a triaxial acceleration matrix sampled at 5 Hz}
  \item{M}{sensor list containing a triaxial magnetometer matrix sampled at 5 Hz}
  \item{P}{sensor list containing a pressure (depth) vector sampled at 5 Hz}
  \item{POS}{sensor list containing a position matrix with columns [sampling time, latitude, longitude]}
}
}
\usage{
harbor_seal
}
\description{
Data is from a _Phoca vitulina_ with tag ID 'hs16_265c'. The device used was a DTAG4 and it was deployed at 2016-09-21 07:55:22 in Husum, Germany.
}
\keyword{datasets}
