% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_nan.R
\name{interp_nan}
\alias{interp_nan}
\title{Remove NAs from sensor data and return indices of (rows of) filled values}
\usage{
interp_nan(data)
}
\arguments{
\item{data}{a data vector or matrix}
}
\value{
A list with entries \code{data} (the input data with NAs filled in) and \code{k} a logical vector indicating the position (if data was a vector) or rows (if data was a matrix) where NAs were filled in. Internal NAs are filled via linear interoplation, while leading and trailing ones are filled using the first following or last preceding good value.
}
\description{
This is an internal function used by \code{\link{htrack}}
}
\examples{
\dontrun{
A <- matrix(c(NA, NA ,3, 4,5,6,7,8,9,10,NA,NA,13,14,15,16, NA, NA), ncol = 2)
result <- interp_nan(A)
}
}
