% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_axis.R
\name{inv_axis}
\alias{inv_axis}
\title{Identify the axis in triaxial movement measurements that varies the least, i.e., the invariant axis.}
\usage{
inv_axis(A)
}
\arguments{
\item{A}{The triaxial sensor measurement axis e.g., from on accelerometer or magnetometer. The frame and unit of A do not matter.}
}
\value{
V A 3x1 vector defining the least varying axis in A. V is a direction vector so has a length of 1 and is unit-less. V is defined in the same frame as A.

q The fraction of movement in the invariant axis. A small number (e.g., <0.05) implies that very little movement occurs in this axis and so the movement is largely planar (i.e., two-dimensional). If the fraction is >> 0.05, the motion in A is better described as three-dimensional. q is a fraction and so does not have a unit.
}
\description{
Rotational and linear movement in some types of propulsion largely occur in 2 dimensions e.g., body rotation in cetacean caudal propulsion occurs around the animal's transverse axis. Likewise sychronized wing flaps in flight or pectoral swimming may generate acceleration in the longitudinal and dorso-ventral axes but much less in the transverse axis. This function identifies the direction of the axis that moves the least in a movement matrix.
}
\note{
This function returns one invariant axis that applies to the entire input signal so it is important to choose a relevant sub-sample of movement data, A, to analyse.
}
