% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_fields.R
\name{merge_fields}
\alias{merge_fields}
\title{Merge the fields of two lists}
\usage{
merge_fields(s1, s2)
}
\arguments{
\item{s1}{Arbitrary list e.g., containing metadata or settings.}

\item{s2}{Arbitrary list e.g., containing metadata or settings.}
}
\value{
A list containing all of the fields in s1 and s2
}
\description{
This function is used to merge the fields of two lists. If there are duplicate fields, the fields in s1 are taken.
}
\examples{
s1 <- list( a = 1, b = c(2,3,4))
         s2 <- list( b = 3, c = 'cat')
	      s <- merge_fields(s1,s2)
	      s #yields list( a = 1, b = c(2,3,4), c = 'cat')
}
