% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_editor.R
\name{metadata_editor}
\alias{metadata_editor}
\title{Edits a html file from given csv.}
\usage{
metadata_editor(
  masterHTML = system.file("extdata", "tagmetadata.html", package = "tagtools"),
  csvfilename
)
}
\arguments{
\item{masterHTML}{default masterHTML is located in the package, or can be changed according to user input.}

\item{csvfilename}{file name of csv to be used for editing the HTML}
}
\value{
A "dynamic tagmetadata.html" which is the masterHTML with changes from csv. This file is written to the current working directory, and also opened for editing by the user.
}
\description{
Takes data from csv, and edits a default or given html to fill in data from the csv. HTML must be tagmetadata.html or variations, csv should only contain metadata of tag.
}
