% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotmat2euler.R
\name{rotmat2euler}
\alias{rotmat2euler}
\title{Decompose a rotation (or direction cosine) matrix into Euler angles, pitch, roll, and heading.}
\usage{
rotmat2euler(Q)
}
\arguments{
\item{Q}{is a 3x3 rotation matrix.}
}
\value{
prh is a 1x3 vector containing: prh=[p,r,h] where p is the pitch angle in radians, r is the roll angle in radians, and h is the heading or yaw angle in radians.
}
\description{
Decompose a rotation (or direction cosine) matrix into Euler angles, pitch, roll, and heading.
}
