% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_wgs84.R
\name{acc_wgs84}
\alias{acc_wgs84}
\title{Calculate total acceleration}
\usage{
acc_wgs84(latitude)
}
\arguments{
\item{latitude}{The latitude in degrees.}
}
\value{
g Given in units of m/s^2
}
\description{
This function calculates the total acceleration due to gravitation and centripetal force at the earth's surface according to the WGS84 internationalgravity formula.
}
\note{
Source: http://solid_earth.ou.edu/notes/potential/igf.htm
}
