% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_cal.R
\name{apply_cal}
\alias{apply_cal}
\title{Implement a calibration on sensor data}
\usage{
apply_cal(X, cal, T)
}
\arguments{
\item{X}{A sensor list or matrix or vector}

\item{cal}{A calibration list for the data in X. For example, this could come from spherical_cal.}

\item{T}{A sensor list or vector of temperature measurements for use in temperature compensation. If T is not a sensor list, it must be the same size and sampling rate as the data in X. T is only required if there is a tcomp field in the cal structure.}
}
\value{
A sensor list with calibration implemented. Data size and sampling rate is the same as for the input data but units may have changed.
}
\description{
Implement a calibration on sensor data
}
\note{
Cal fields currently supported are : poly, cross, map, tcomp, tref
}
\examples{
\dontrun {#Will come soon!}
}
