% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_line.R
\name{col_line}
\alias{col_line}
\title{Plot coloured line(s) in 2 dimensions in the current figure.}
\usage{
col_line(x, y, c)
}
\arguments{
\item{x}{A vector or matrix of points on the horizontal axis.}

\item{y}{A vector or matrix of points on the vertical axis.}

\item{c}{A vector or matrix of values representing the colour to draw at each point.}
}
\description{
Plot coloured line(s) in 2 dimensions in the current figure.
}
\note{
x, y and c must all be the same size. If x, y, and c are matrices, one line is drawn for each column. The color axis will by default span the range of values in c, i.e., caxis will be c(min(min(c)), max(max(c))). This can be changed by calling caxis after colline.
}
