% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_line.R
\name{col_line}
\alias{col_line}
\title{Plot coloured line(s) in 2 dimensions}
\usage{
col_line(formula, color, data = NULL, ...)
}
\arguments{
\item{formula}{formula of the form y ~ x where y and x are the (unquoted) names of the variables to put on the x and y axes of the plot}

\item{color}{one-sided formula of the form ~ c giving the (unquoted) name of the variable by which to color. Ideally c should be a factor or character variable.}

\item{data}{data.frame or tibble in which variables x, y, and c are found.}

\item{...}{Additional inputs to be passed to gf_path()}
}
\value{
a ggplot object with the requested plot
#' @export
}
\description{
This function is used to plot two dimensional lines with each individual line possessing a different color.
}
