% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_line.R
\name{col_line}
\alias{col_line}
\title{Plot coloured line(s) in 2 dimensions}
\usage{
col_line(
  formula,
  x = NULL,
  y = NULL,
  c,
  c_lab = quote(c),
  data = NULL,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{formula}{a formula of the form y ~ x giving the (unquoted) variable names to plot}

\item{x}{ignored if a formula is provided. A vector or matrix of points on the horizontal axis.}

\item{y}{ignored if a formula is provided. A vector or matrix of points on the vertical axis.}

\item{c}{A vector or matrix of values representing the colour to draw at each point.}

\item{c_lab}{A string to use as the label for the color legend}

\item{data}{(optional) a data.frame containing variables used for plotting}

\item{interactive}{logical. Plot interactive or static figure? Note: For some reason it is much faster to plot a static figure and then call ggplotly() outside this function, e.g., F <- col_line(y~x, c = z); ggplotly(F)}

\item{...}{Additional inputs to be passed to gf_path()}
}
\value{
If output is assigned to an object, it will be a ggplot (or ggplotly) object and no plot will be displayed. Otherwise, the plot will be rendered.
}
\description{
This function is used to plot two dimensional lines with each individual line possessing a different color.
}
\note{
x, y and c must all be vectors of the same size.
}
