% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_fields.R
\name{merge_fields}
\alias{merge_fields}
\title{Merge the fields of two structures. If there are duplicate fields, the fields in s1 are taken.}
\usage{
merge_fields(s1, s2)
}
\arguments{
\item{s1}{Arbitrary list e.g., containing metadata or settings.}

\item{s2}{Arbitrary list e.g., containing metadata or settings.}
}
\value{
s A list containing all of the fields in s1 and s2
}
\description{
Merge the fields of two structures. If there are duplicate fields, the fields in s1 are taken.
}
