% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrack.R
\name{ptrack}
\alias{ptrack}
\title{Estimate simple dead-reckoned track}
\usage{
ptrack(A, M, s, sampling_rate, fc = NULL, include_pe = NULL)
}
\arguments{
\item{A}{An nx3 acceleration matrix with columns [ax ay az] or acceleration sensor list (e.g., from readtag.R). Acceleration can be in any consistent unit, e.g., g or m/s^2.}

\item{M}{The magnetometer signal matrix, M = [mx,my,mz] in any consistent unit (e.g., in uT or Gauss)  or magnetometer sensor list (e.g., from readtag.R). A and M must have the same size (and so are both measured at the same sampling rate).}

\item{s}{The forward speed of the animal in m/s. s can be a single number meaning that the animal is assumed to travel at a constant speed. s can also be a vector with the same number of rows as A and M, e.g., generated by speed_from_depth().}

\item{sampling_rate}{The sampling rate of the sensor data in Hz (samples per second).}

\item{fc}{(optional) The cut-off frequency of a low-pass filter to aply to A and M before computing bodypointing angle. The filter cut-off frequency is in Hz. The filter length is 4*sampling_rate/fc. Filtering adds no group delay. If fc is empty or not given, the default value of 0.2 Hz (i.e., a 5 second time constant) is used.}

\item{include_pe}{A conditional statement. If include_pe is TRUE, the estimated depth of altitude predicted will be returned with the estimated track.}
}
\value{
The estimated track in a local level frame. The track is defined as meters of northward and eastward movement (termed 'northing' and 'easting', i.e, T=[northing,easting]) relative to the animal'sposition at the start of the measurements  (which is defined as [0,0]). The track sampling rate is the same as for the input data and so each row of T defines the track coordinates at times 0,1/sampling_rate,2/sampling_rate,... relative to the start time of the measurements.

OR, if include_pe = TRUE, this function returns the above value and the estimated depth or altitude predicted from the speed and pitch angle. This can be compared against the measured depth/altitude to assess errors in the dead-reckoned track. Note that even if pe matches the observed depth, this does not guarantee that the track is accurate.
}
\description{
This function is used to estimate the simple dead-reckoned track (pseudo-track) based on speed and bodypointing angle.
}
\details{
Possible input combinations: ptrack(A,M,s,sampling_rate) if A and M are matrices, ptrack(A,M,s,sampling_rate,fc) if A and M are matrices, ptrack(A,M,s) if A and M are lists, ptrack(A,M,s,fc = fc) if A and M are lists.
}
\note{
Frame: This function assumes a [north,east,up] navigation frame and a [forward,right,up] local frame. Both A and M must be rotated if needed to match the animal's cardinal axes otherwise the track will not be meaningful.

CAUTION: dead-reckoned tracks are usually very inaccurate. They are useful to get an idea of HOW animals move rather than WHERE they go. Few animals probably travel in exactly the direction of their longitudinal axis and anyway measuring the precise orientation of the longitudinal axis of a non-rigid animal is fraught with error. Moreover, if there is net flow in the medium, the animal will be affected by the flow in addition to its autonomous movement. For swimming animals this can lead to substantial errors. The forward speed is assumed to be with respect to the medium so the track derived here is NOT the 'track-made-good', i.e., the geographic movement of the animal. It estimates the movement of the animal with respect to the medium. There are numerous other sources of error so use at your own risk!
}
