% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_nc.R
\name{save_nc}
\alias{save_nc}
\title{Save a tag dataset to a netCDF file.}
\usage{
save_nc(file, X, ...)
}
\arguments{
\item{file}{The name of the data and metadata file to be written. If \code{file} does not include a .nc suffix, this will be added automatically.}

\item{X}{An \code{animaltag} object, or a list of tag sensor and/or metadata lists. Alternatively, sensor and metadata lists may be input as multiple separate unnamed inputs. Only these kind of variables can be saved
in a NetCDF file because the supporting information in these structures is
needed to describe the contents of the file. For non-archive and non-portable
storage of variables, consider using \code{\link{save}} or various functions to write data to text files.}

\item{...}{Additional sensor or metadata lists, if user has not bundled them all into a list already but is providing individual structures.}
}
\description{
This function saves a tag dataset to a netCDF file (this is an archival file format supported by the tagtools package and suitable for submission to online data archives).
}
\details{
Warning: this will overwrite any previous NetCDF file with the same name. The file is assumed to be in the current working directory unless \code{file} includes file path information.
}
\examples{
\dontrun{
save_nc("dog17_124a", A, M, P, info)
# or equivalently:
save_nc("dog17_124a", X = list(A, M, P, info))
# generates a file dog17_124a.nc and adds variables A, M and P, and a metadata structure.
}
}
